requirements.txt
mljar-mercury
pandas
keras
tensorflow-cpu
sklearn